BukGet Server Initial Release 0.01
Initial release of the bukget server cronjob.  This script is designed to
perform most of the maintenance functions necessary to keep the
repository up to date and to keep the bukkit packages themselves up
to date as well.  Some functionality is currently missing however (like
the ability to remove orphaned packages) however that will be added
in a future release.