Added :version and download links
Added the ability to only pull a specific version from the database
based on the name of the version.  There is also a alias "latest"
that will always pull the most current plugin version based on
date.

/api/plugin/:plugin_name/:version

Also added the ability to simply call the version as before however
with a /download at the end to redirect to the download link of
that specific plugin.

/api/plugin/:plugin_name/:version/download