Added update_sqlite() and integrated into update function.
update_sqlite generates a sqlite database containing all of the plugin data as well as all of the version information for each plugin.  This was created as a way to minimize download and processing times as it was discovered that parsing /api/json was taking too much time and locking up servers.