Version Tracking
This will allow a requesting script to track a hash or filename that was searched, with an internal list. This way the requesting script can track multiple files on local disk to replace and update with a single request.

This change will also report the given version of the requested hash, if given. This allows the requester to know that version X.XX needs to be upgraded to version Y.YY.